Start moving notebook-centric routines into their own files
