+++ moldesign/molecules/molecule.py
@@ -12,15 +12,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import itertools
+
 import numpy as np
 
 import moldesign as mdt
 from moldesign import helpers, utils
-from moldesign.exceptions import NotCalculatedError
 from moldesign import units as u
 from moldesign.compute import DummyJob
+from moldesign.exceptions import NotCalculatedError
 from moldesign.min.base import MinimizerBase
-
+from moldesign.molecules.notebook_display import MolNotebookMixin
 from . import toplevel, Residue, Chain, Instance, AtomGroup, Bond
 from .coord_arrays import *
 
@@ -389,121 +390,6 @@ class MolDrawingMixin(object):
         if 'wfn' not in self.properties:
             self.calculate_wfn()
         return OrbitalViewer(self, **kwargs)
-
-
-class MolReprMixin(object):
-    """ Methods for creating text-based representations of the molecule
-
-    Note:
-        This is a mixin class designed only to be mixed into the :class:`Molecule` class. Routines
-        are separated are here for code organization only - they could be included in the main
-        Molecule class without changing any functionality
-    """
-    def __repr__(self):
-        try:
-            return '<%s (%s), %d atoms>' % (self.name,
-                                            self.__class__.__name__,
-                                            len(self.atoms))
-        except:
-            return '<molecule (error in __repr__) at %s>' % id(self)
-
-    def __str__(self):
-        return 'Molecule: %s' % self.name
-
-    def markdown_summary(self):
-        """A markdown description of this molecule.
-
-        Returns:
-            str: Markdown"""
-        # TODO: remove leading underscores for descriptor-protected attributes
-        lines = ['### Molecule: "%s" (%d atoms)' % (self.name, self.natoms)]
-
-        if self.description:
-            lines.append(self.description[:5000])
-
-        lines.extend([
-                 '**Mass**: {:.2f}'.format(self.mass),
-                 '**Formula**: %s' % self.get_stoichiometry(html=True),
-                 '**Charge**: %s'%self.charge])
-
-        if self.energy_model:
-            lines.append('**Potential model**: %s' % str(self.energy_model))
-
-        if self.integrator:
-            lines.append('**Integrator**: %s' % str(self.integrator))
-
-        if self.is_biomolecule:
-            lines.extend(self.biomol_summary_markdown())
-
-        return '\n\n'.join(lines)
-
-    def _repr_markdown_(self):
-        return self.markdown_summary()
-
-    def biomol_summary_markdown(self):
-        """A markdown description of biomolecular structure.
-
-        Returns:
-            str: Markdown string"""
-        lines = []
-        if len(self.residues) > 1:
-            table = self.get_residue_table()
-            lines.append('### Residues')
-            # extra '|' here may be workaround for a bug in ipy.markdown?
-            lines.append(table.markdown(replace={0: ' '}) + '|')
-
-            lines.append('### Biopolymer chains')
-            seqs = []
-            for chain in self.chains:
-                seq = chain.sequence
-                if not seq.strip():  # don't write anything if there's no sequence
-                    continue
-
-                # deal with extra-long sequences
-                seqstring = []
-                for i in xrange(0, len(seq), 80):
-                    seqstring.append(seq[i:i + 80])
-                seqstring = '\n'.join(seqstring)
-                seqs.append('**%s**: `%s`' % (chain.name, seqstring))
-            lines.append('<br>'.join(seqs))
-
-        return lines
-
-    def get_residue_table(self):
-        """Creates a data table summarizing this molecule's primary structure.
-
-        Returns:
-            moldesign.utils.MarkdownTable"""
-        table = utils.MarkdownTable(*(['chain'] +
-                                      'protein dna rna unknown water solvent'.split()))
-        for chain in self.chains:
-            counts = {}
-            unk = []
-            for residue in chain.residues:
-                cat = residue.type
-                if cat == 'unknown':
-                    unk.append(residue.name)
-                counts[cat] = counts.get(cat, 0) + 1
-            counts['chain'] = '<pre><b>%s</b></pre>' % chain.name
-            if 0 < len(unk) <= 4:
-                counts['unknown'] = ','.join(unk)
-            table.add_line(counts)
-        return table
-
-    def get_stoichiometry(self, html=False):
-        """ Return this molecule's stoichiometry
-
-        Returns:
-            str
-        """
-        counts = {}
-        for atom in self.atoms:
-            counts[atom.symbol] = counts.get(atom.symbol, 0) + 1
-
-        my_elements = sorted(counts.keys())
-        if html: template = '%s<sub>%d</sub>'
-        else: template = '%s%d'
-        return ''.join([template % (k, counts[k]) for k in my_elements])
 
 
 class MolTopologyMixin(object):
@@ -894,7 +780,7 @@ class Molecule(AtomGroup,
                MolConstraintMixin,
                MolPropertyMixin,
                MolDrawingMixin,
-               MolReprMixin,
+               MolNotebookMixin,
                MolTopologyMixin, MolSimulationMixin):
     """
     ``Molecule`` objects store a molecular system, including atoms, 3D coordinates, molecular
@@ -982,7 +868,7 @@ class Molecule(AtomGroup,
             - :class:`moldesign.molecules.MolSimulationMixin`
             - :class:`moldesign.molecules.MolTopologyMixin`
             - :class:`moldesign.molecules.MolConstraintMixin`
-            - :class:`moldesign.molecules.MolReprMixin`
+            - :class:`moldesign.molecules.MolNotebookMixin`
     """
     positions = ProtectedArray('_positions')
     momenta = ProtectedArray('_momenta')
@@ -993,14 +879,13 @@ class Molecule(AtomGroup,
                  pdbname=None,
                  charge=None,
                  electronic_state_index=0,
-                 description=None):
+                 metadata=None):
         super(Molecule, self).__init__()
 
         # copy atoms from another object (i.e., a molecule)
         oldatoms = helpers.get_all_atoms(atomcontainer)
 
         if copy_atoms or (oldatoms[0].molecule is not None):
-            #print 'INFO: Copying atoms into new molecule'
             atoms = oldatoms.copy()
             if name is None:  # Figure out a reasonable name
                 if oldatoms[0].molecule is not None:
@@ -1021,7 +906,7 @@ class Molecule(AtomGroup,
         self.constraints = utils.ExclusiveList(key=utils.methodcaller('_constraintsig'))
         self.energy_model = None
         self.integrator = None
-        self.description = description
+        self.metadata = utils.if_not_none(metadata, utils.DotDict())
         self.electronic_state_index = electronic_state_index
 
         if charge is not None:
@@ -1045,6 +930,17 @@ class Molecule(AtomGroup,
 
         self._properties = MolecularProperties(self)
         self.ff = utils.DotDict()
+
+    def __repr__(self):
+        try:
+            return '<%s (%s), %d atoms>' % (self.name,
+                                            self.__class__.__name__,
+                                            len(self.atoms))
+        except:
+            return '<molecule (error in __repr__) at %s>' % id(self)
+
+    def __str__(self):
+        return 'Molecule: %s' % self.name
 
     def newbond(self, a1, a2, order):
         """ Create a new bond
